refactor(home): remove unnecessary folders
