--- apps/frontend/src/pages/_home/todays-pick/TodaysPickHeader.vue
@@ -1,19 +1,0 @@
-<script setup lang="ts"></script>
-
-<template>
-  <p
-    class="text-terracotta mb-3 text-sm font-semibold tracking-widest uppercase"
-  >
-    Today's Pick
-  </p>
-  <h2
-    class="font-display text-4xl font-bold tracking-tight text-white lg:text-5xl"
-  >
-    Culinary Calendar
-  </h2>
-  <p class="mt-6 max-w-lg text-lg leading-relaxed text-stone-300">
-    Every day we pick seasonal recipes that perfectly match the time of year.
-    Today — a silky roasted pumpkin soup with aromatic spices and homemade
-    croutons.
-  </p>
-</template>
